+++ Image-Based Post-Processing Effects/Engine/Code/Graphics/OpenGL/EnvironmentRenderer.cpp
@@ -126,7 +126,7 @@ void EnvironmentRenderer::calculateReflectance(const std::shared_ptr<Environment
 
 		float roughness = (float)mip / (float)(maxMipLevels - 1);
 		uRoughnessRO.set(roughness * roughness);
-		uImageSizeRO.set(glm::vec2(mipRes));
+		uImageSizeRO.set(glm::vec2(static_cast<float>(mipRes)));
 		glBindImageTexture(0, _environmentProbe->getReflectionTexture()->getId(), mip, GL_FALSE, 0, GL_WRITE_ONLY, GL_R11F_G11F_B10F);
 		glDispatchCompute(mipRes / 8, mipRes / 8, 1);
 	}
